[19] move files around for refactor
