+++ BodyPositiveZumba/Common/WebView/WebView.swift
@@ -1,0 +1,22 @@
+//
+//  WebView.swift
+//  Common
+//
+//  Created by Andrew Elbert on 2/1/25.
+//
+
+import SwiftUI
+import WebKit
+
+struct WebView: UIViewRepresentable {
+    let url: URL
+
+    func makeUIView(context: Context) -> WKWebView {
+        return WKWebView()
+    }
+
+    func updateUIView(_ uiView: WKWebView, context: Context) {
+        let request = URLRequest(url: url)
+        uiView.load(request)
+    }
+}
