+++ BodyPositiveZumba/Common/WebView/WebViewURLModel.swift
@@ -1,0 +1,13 @@
+//
+//  WebViewURLModel.swift
+//  Common
+//
+//  Created by Andrew Elbert on 2/1/25.
+//
+
+import Foundation
+
+struct WebViewURL: Identifiable {
+    let id = UUID()
+    let url: URL
+}
