+++ BodyPositiveZumba/Massage/MassageUI/MassageView.swift
@@ -6,15 +6,9 @@
 //
 
 import SwiftUI
-import UIKit
-import WebKit
-
-struct BookingURL: Identifiable {
-    let id = UUID()
-    let url: URL
-}
 
 struct MassageView: View {
+    
     let cardsData: [CardModel] = [
         CardModel(
             id: UUID(),
@@ -36,7 +30,7 @@ struct MassageView: View {
     @State private var currentIndex: Int = 0
     @State private var isAnimating = false
     @State private var isCardExpanded = false
-    @State private var bookingURL: BookingURL?
+    @State private var bookingURL: WebViewURL?
 
     var body: some View {
         ZStack(alignment: .topTrailing) {
@@ -44,11 +38,12 @@ struct MassageView: View {
                 Text(Constants.Massage.pageTitle)
                     .font(.system(size: 34, weight: .bold, design: .serif))
                     .multilineTextAlignment(.center)
+                    .foregroundColor(Constants.Colors.darkerCyan)
                     .padding()
 
                 Text(Constants.Massage.pageBio)
                     .font(.system(size: 16, design: .serif))
-                    .foregroundColor(.gray)
+                    .foregroundColor(.black.opacity(0.7))
                     .multilineTextAlignment(.center)
                     .padding(.horizontal)
                     .padding(.bottom, 16)
@@ -79,15 +74,15 @@ struct MassageView: View {
                 Button(action: {
                     let normalizedIndex = ((currentIndex % cardsData.count) + cardsData.count) % cardsData.count
 
-                    var urlString: String = ""
+                    var urlString: String
                     if normalizedIndex == 0 {
-                        urlString = "https://ladyloveholistics.com/"
+                        urlString = Constants.Massage.ladyLoveHolisticURL
                     } else {
-                        urlString = "https://swanns-healing-elements.square.site/"
+                        urlString = Constants.Massage.swannsHealingElementsURL
                     }
 
                     if let url = URL(string: urlString) {
-                        bookingURL = BookingURL(url: url)
+                        bookingURL = WebViewURL(url: url)
                     }
                 }) {
                     Text("Book Today!")
@@ -125,119 +120,5 @@ struct MassageView: View {
                     }))
                 .edgesIgnoringSafeArea(.all)
         }
-    }
-}
-
-struct SwipeAnimationView: View {
-    @Binding var isAnimating: Bool
-
-    var body: some View {
-        HStack(spacing: 16) {
-            ZStack {
-                Image(systemName: "hand.point.left.fill")
-                    .offset(x: isAnimating ? -10 : 0)
-                    .opacity(isAnimating ? 0 : 1)
-
-                Image(systemName: "arrow.left")
-                    .offset(x: isAnimating ? -10 : 0)
-                    .opacity(isAnimating ? 1 : 0)
-            }
-
-            ZStack {
-                Image(systemName: "hand.point.right.fill")
-                    .offset(x: isAnimating ? 10 : 0)
-                    .opacity(isAnimating ? 0 : 1)
-
-                Image(systemName: "arrow.right")
-                    .offset(x: isAnimating ? 10 : 0)
-                    .opacity(isAnimating ? 1 : 0)
-            }
-        }
-        .font(.system(size: 24))
-        .foregroundColor(.gray)
-        .frame(width: 120, height: 30)
-        .onAppear {
-            withAnimation(Animation.easeInOut(duration: 1.2).repeatForever(autoreverses: true)) {
-                isAnimating = true
-            }
-        }
-    }
-}
-
-struct InfiniteCarouselView<Content: View, T: Identifiable>: View {
-    var items: [T]
-    @Binding var currentIndex: Int
-    var spacing: CGFloat = 10
-    var sideSpacing: CGFloat = 40
-    var content: (T, Bool) -> Content
-
-    private var circularItems: [T] {
-        guard let first = items.first, let last = items.last else { return items }
-        return [last] + items + [first]
-    }
-
-    @GestureState private var dragOffset: CGFloat = 0
-
-    var body: some View {
-        GeometryReader { proxy in
-            let cardWidth = max(0, proxy.size.width - (sideSpacing * 2))
-            let cardWithSpacing = cardWidth + spacing
-            let displayIndex = CGFloat(currentIndex + 1)
-
-            HStack(spacing: spacing) {
-                ForEach(Array(circularItems.enumerated()), id: \.offset) { index, item in
-                    content(item, index == currentIndex + 1)
-                        .frame(width: cardWidth)
-                }
-            }
-            .offset(x: sideSpacing - (displayIndex * cardWithSpacing) + dragOffset)
-            .animation(.easeInOut, value: dragOffset)
-            .gesture(
-                DragGesture()
-                    .updating($dragOffset) { value, state, _ in
-                        state = value.translation.width
-                    }
-                    .onEnded { value in
-                        let threshold: CGFloat = cardWidth / 2
-                        var newIndex = currentIndex
-                        if value.translation.width < -threshold {
-                            newIndex += 1
-                        } else if value.translation.width > threshold {
-                            newIndex -= 1
-                        }
-                        withAnimation(.easeInOut) {
-                            currentIndex = newIndex
-                        }
-                    }
-            )
-            .onChange(of: currentIndex) { _, newValue in
-                if newValue < 0 {
-                    DispatchQueue.main.async {
-                        withAnimation(.none) {
-                            currentIndex = items.count - 1
-                        }
-                    }
-                } else if newValue >= items.count {
-                    DispatchQueue.main.async {
-                        withAnimation(.none) {
-                            currentIndex = 0
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-struct WebView: UIViewRepresentable {
-    let url: URL
-
-    func makeUIView(context: Context) -> WKWebView {
-        return WKWebView()
-    }
-
-    func updateUIView(_ uiView: WKWebView, context: Context) {
-        let request = URLRequest(url: url)
-        uiView.load(request)
     }
 }
